Rewrites JSON methods for parse factory.
