+++ ts/parse.ts
@@ -28,7 +28,7 @@ import {Menu} from './menu.js';
 import {MenuUtil} from './menu_util.js';
 import {ContextMenu} from './context_menu.js';
 import {Variable} from './variable.js';
-import {VariablePool} from './variable_pool.js';
+// import {VariablePool} from './variable_pool.js';
 import {Checkbox} from './item_checkbox.js';
 import {Combo} from './item_combo.js';
 import {Label} from './item_label.js';
@@ -38,104 +38,19 @@ import {Rule} from './item_rule.js';
 import {Item} from './item.js';
 import {Slider} from './item_slider.js';
 import {SubMenu} from './sub_menu.js';
-import {SelectionMenu, SelectionBox, SelectionOrder} from './selection_box.js';
+import {SelectionMenu, SelectionBox} from './selection_box.js';
 import {ParserFactory} from './parser_factory.js';
 
 
 export namespace Parse {
 
-  /**
-   * Parses a JSON respresentation of a command item.
-   * @param {JSON} json The JSON object to parse.
-   * @param {Menu} menu The menu the item is attached to.
-   * @return {Command} The new command object.
-   */
-  const command = function(
-    {content: content, action: action, id: id}:
-    {content: string, action: Function, id: string}, menu: Menu): Command {
-    return new Command(menu, content, action, id);
-  };
   ParserFactory.add('command', Command.fromJson.bind(Command));
-
-  /**
-   * Parses a JSON respresentation of a checkbox item.
-   * @param {JSON} json The JSON object to parse.
-   * @param {Menu} menu The menu the item is attached to.
-   * @return {Checkbox} The new checkbox object.
-   */
-  const checkbox = function(
-    {content: content, variable: variable, id: id}:
-    {content: string, variable: string, id: string}, menu: Menu): Checkbox {
-    return new Checkbox(menu, content, variable, id);
-  };
   ParserFactory.add('checkbox', Checkbox.fromJson.bind(Checkbox));
-
-  /**
-   * Parses a JSON respresentation of a combo item.
-   * @param {JSON} json The JSON object to parse.
-   * @param {Menu} menu The menu the item is attached to.
-   * @return {Combo} The new combo object.
-   */
-  const combo = function(
-    {content: content, variable: variable, id: id}:
-    {content: string, variable: string, id: string}, menu: Menu): Combo {
-    return new Combo(menu, content, variable, id);
-  };
   ParserFactory.add('combo', Combo.fromJson.bind(Combo));
-
-  /**
-   * Parses a JSON respresentation of a slider item.
-   * @param {JSON} json The JSON object to parse.
-   * @param {Menu} menu The menu the item is attached to.
-   * @return {Slider} The new combo object.
-   */
-  const slider = function(
-    {content: content, variable: variable, id: id}:
-    {content: string, variable: string, id: string}, menu: Menu): Slider {
-    return new Slider(menu, content, variable, id);
-  };
   ParserFactory.add('slider', Slider.fromJson.bind(Slider));
-
-  /**
-   * Parses a JSON respresentation of a label item.
-   * @param {JSON} json The JSON object to parse.
-   * @param {Menu} menu The menu the item is attached to.
-   * @return {Label} The new label object.
-   */
-  const label = function(
-    {content: content, id: id}: {content: string, id: string},
-    menu: Menu): Label {
-    return new Label(menu, content, id);
-  };
   ParserFactory.add('label', Label.fromJson.bind(Label));
-
-  /**
-   * Parses a JSON respresentation of a radio item.
-   * @param {JSON} json The JSON object to parse.
-   * @param {Menu} menu The menu the item is attached to.
-   * @return {Radio} The new radio object.
-   */
-  export const radio = function(
-    {content: content, variable: variable, id: id}:
-    {content: string, variable: string, id: string}, menu: Menu): Radio {
-    return new Radio(menu, content, variable, id);
-  };
   ParserFactory.add('radio', Radio.fromJson.bind(Radio));
-
-  /**
-   * Parses a JSON respresentation of a submenu item.
-   * @param {JSON} json The JSON object to parse.
-   * @param {Menu} menu The menu the item is attached to.
-   * @return {Submenu} The new submenu object.
-   */
-  const submenu = function(
-    {content: content, menu: submenu, id: id}:
-    {content: string, menu: any, id: string}, menu: Menu): Submenu {
-    const item = new Submenu(menu, content, id);
-    const sm = subMenu(submenu, item);
-    item.submenu = sm;
-    return item;
-  };
+  ParserFactory.add('rule', Rule.fromJson.bind(Rule));
   ParserFactory.add('submenu', Submenu.fromJson.bind(Submenu));
 
   /**
@@ -154,36 +69,7 @@ export namespace Parse {
     return ParserFactory.get('contextMenu')({menu: menu});
   };
   ParserFactory.add('contextMenu', ContextMenu.fromJson.bind(ContextMenu));
-
-  /**
-   * Parses a JSON respresentation of a submenu.
-   * @param {JSON} json The JSON object to parse.
-   * @param {Submenu} anchor The anchor item the submenu is attached to.
-   * @return {SubMenu} The new submenu object.
-   */
-  export const subMenu = function(
-    {items: its}: {items: any[], id: string},
-    anchor: Submenu): SubMenu {
-    const submenu = new SubMenu(anchor);
-    const itemList = items(its, submenu);
-    submenu.items = itemList;
-    return submenu;
-  };
   ParserFactory.add('subMenu', SubMenu.fromJson.bind(SubMenu));
-
-  /**
-   * Parses a JSON respresentation of a variable and inserts it into the
-   * variable pool of the context menu.
-   * @param {JSON} json The JSON object to parse.
-   */
-  export const variable = function(
-    {name: name, getter: getter, setter: setter}:
-    {name: string, getter: () => string | boolean,
-     setter: (x: (string | boolean)) => void},
-    pool: VariablePool<string|boolean>) {
-    const variable = new Variable(name, getter, setter);
-    pool.insert(variable);
-  };
   ParserFactory.add('variable', Variable.fromJson.bind(Variable));
 
   /**
@@ -196,7 +82,7 @@ export namespace Parse {
     return hidden.map(x => x[0]);
   };
   ParserFactory.add('items', items);
-  
+
   /**
    * Parses items in JSON formats and attaches them to the menu.
    * @param {Array.<JSON>} items List of JSON menu items.
@@ -204,7 +90,6 @@ export namespace Parse {
    * @return {Item} The list of parsed items.
    */
   const item = function(item: any, ctxt: Menu): Item {
-    // const func = mapping[item['type']];
     const func = ParserFactory.get(item['type']);
     if (func) {
       const menuItem = func(item, ctxt);
@@ -217,54 +102,13 @@ export namespace Parse {
     return null;
   };
 
-  // /**
-  //  * @type{Object.<Function>}
-  //  */
-  // const mapping: { [id: string]: Function } = {
-  //   'checkbox': ParserFactory.get('checkbox'),
-  //   'combo': ParserFactory.get('combo'),
-  //   'command': ParserFactory.get('command'),
-  //   'label': ParserFactory.get('label'),
-  //   'radio': ParserFactory.get('radio'),
-  //   'rule': ParserFactory.get('rule'),
-  //   'slider': ParserFactory.get('slider'),
-  //   'submenu': ParserFactory.get('submenu')
-  // };
+  ParserFactory.add('selectionMenu',
+                    SelectionMenu.fromJson.bind(SelectionMenu));
+  ParserFactory.add('selectionBox', SelectionBox.fromJson.bind(SelectionBox));
 
-
-  declare type selection = {title: string, values: string[], variable: string};
-  
-  const selectionMenu = function(
-    {title: title, values: values, variable: variable}: selection,
-    sb: SelectionBox): SelectionMenu {
-    let selection = new SelectionMenu(sb);
-    let tit = Label.fromJson({content: title || '', id: title || 'id'}, selection);
-    let rul = Rule.fromJson({}, selection);
-    let radios = values.map(x => radio({content: x, variable: variable, id: x}, selection));
-    let items = [tit, rul].concat(radios) as Item[];
-    selection.items = items;
-    return selection;
-  };
-  
   export const selectionBox = function(
-    {title: title, signature: signature, selections: selections, order: order}:
-    {title: string, signature: string, selections: selection[], order?: SelectionOrder},
-    ctxt: ContextMenu): SelectionBox {
-    let sb = new SelectionBox(title, signature, order);
-    sb.attachMenu(ctxt);
-    let sels = selections.map(x => selectionMenu(x, sb));
-    sb.selections = sels;
-    return sb;
+    json: JSON, ctxt: ContextMenu): SelectionBox {
+    return ParserFactory.get('selectionBox')(json, ctxt);
   };
 
-  export const dummy = function() {
-    console.log(command);
-    console.log(combo);
-    console.log(checkbox);
-    console.log(radio);
-    console.log(slider);
-    console.log(submenu);
-    console.log(label);
-  };
 }
-
